requirements_base.txt
otree-core>=1.4.12
Django==1.11.10 # for heroku, needs to be explicitly in requirements file
